Move files to `components/History/Multiple`
