+++ client/src/components/History/Multiple/MultipleView.vue
